requirements.txt
pytesseract
Pillow

